sdk-go: Add test for ExponentialBackoff & fix types

+++ sdk-go/littlehorse/wf_lib_internal.go
@@ -774,7 +774,7 @@ func (tn *TaskNodeOutput) withExponentialBackoffImpl(policy *lhproto.Exponential
 	return tn
 }
 
-func (tn *TaskNodeOutput) withRetriesImpl(retries int) *TaskNodeOutput {
+func (tn *TaskNodeOutput) withRetriesImpl(retries int32) *TaskNodeOutput {
 	tn.parent.overrideTaskRetries(tn, retries)
 	return tn
 }
@@ -1032,7 +1032,7 @@ func (c *WorkflowCondition) getReverse() *lhproto.EdgeCondition {
 	return out
 }
 
-func (t *WorkflowThread) overrideTaskRetries(taskNodeOutput *TaskNodeOutput, retries int) {
+func (t *WorkflowThread) overrideTaskRetries(taskNodeOutput *TaskNodeOutput, retries int32) {
 	t.checkIfIsActive()
 
 	node := t.spec.Nodes[taskNodeOutput.Output.nodeName]
@@ -1041,7 +1041,7 @@ func (t *WorkflowThread) overrideTaskRetries(taskNodeOutput *TaskNodeOutput, ret
 		t.throwError(errors.New("impossible to not have task node here"))
 	}
 
-	node.GetTask().Retries = int32(retries)
+	node.GetTask().Retries = retries
 }
 
 func (t *WorkflowThread) overrideTaskExponentialBackoffPolicy(taskNodeOutput *TaskNodeOutput, policy *lhproto.ExponentialBackoffRetryPolicy) {
